+++ src/traits/to_u64.rs
@@ -102,7 +102,7 @@ mod tests {
                 2, 4, 8, 16, 32, 64, 64, 256,
                 u16::MAX as u64,
                 u32::MAX as u64,
-                u64::MAX as u64,
+                u64::MAX,
             ];
 
             for &value in VALUES {
